extracted publisher class
